AGENT-222: Unstick queue if we get RESOURCE_LOCKED
If a queue is marked as durable and the client abruptly disconnects from the
broker, the queue will hang around and prevent the client from re-attaching
once it's back. When this condition is detected, send a message to the queue
(forcing rabbitmq to update its state) and then reconnect the client to the
broker.